+++ embed/CHANGELOG.md
@@ -7,7 +7,7 @@
   instead of the package root directory in a pub workspace ([#55])
 
 [#53]: https://github.com/fujidaiti/embed.dart/pull/53
-[#53-guide]: https://github.com/fujidaiti/embed.dart/blob/master/embed/README.md#i-edited-my-json-file-to-embed-but-the-generated-code-doesnt-update-even-when-i-run-build_runner-again
+[#53-guide]: https://github.com/fujidaiti/embed.dart/blob/master/embed/README.md#which-files-can-be-tracked
 [#55]: https://github.com/fujidaiti/embed.dart/issues/55
 
 ## 1.6.7